Address PR comments.

+++ src/sagemaker_containers/_intermediate_output.py
@@ -44,17 +44,15 @@ def _upload_to_s3(s3_uploader, relative_path, file_path, filename):
     try:
         key = os.path.join(s3_uploader['key_prefix'], relative_path, filename)
         s3_uploader['transfer'].upload_file(file_path, s3_uploader['bucket'], key)
-
-        # delete the original file
-        os.remove(file_path)
     except FileNotFoundError:
         # Broken link or deleted
         pass
     except Exception:
         logger.exception('Failed to upload file to s3.')
-
+    finally:
         # delete the original file
-        os.remove(file_path)
+        if os.path.exists(file_path):
+            os.remove(file_path)
 
 
 def _copy_file(executor, s3_uploader, relative_path, filename):
